Renames rswag-api to open_api-rswag-api
