--- rswag-api/lib/rswag/api.rb
@@ -1,14 +1,0 @@
-require 'rswag/api/configuration'
-require 'rswag/api/engine'
-
-module Rswag
-  module Api
-    def self.configure
-      yield(config)
-    end
-
-    def self.config
-      @config ||= Configuration.new
-    end
-  end
-end
