+++ rswag-api/open_api-rswag-api.gemspec
@@ -2,9 +2,9 @@ $:.push File.expand_path("../lib", __FILE__)
 
 # Describe your gem and declare its dependencies:
 Gem::Specification.new do |s|
-  s.name        = "rswag-api"
+  s.name        = "open_api-rswag-api"
   s.version     = ENV['TRAVIS_TAG'] || '0.0.0'
-  s.authors     = ["Richie Morris"]
+  s.authors     = ["Richie Morris", "Jay Danielian"]
   s.email       = ["domaindrivendev@gmail.com"]
   s.homepage    = "https://github.com/domaindrivendev/rswag"
   s.summary     = "A Rails Engine that exposes Swagger files as JSON endpoints"
