+++ rswag-api/spec/open_api/rswag/api/fixtures/config/routes.rb
