+++ rswag-api/spec/open_api/rswag/api/fixtures/swagger/v1/swagger.json
